Add files via upload
I believe that the significant increase in short contigs is due to the trimming creating much smaller reads that failed to overlap as well as the previously un-trimmed version. This also increased the number of contigs about 8 fold. The lack of overlapping reads --> lack of overlapping contigs --> more contigs --> much smaller "largest contig" --> lower N50 overall.